+++ R/utils.R
@@ -12,6 +12,10 @@ isNumber = function(x, minimum = NA, maximum = NA) {
            (is.na(maximum) || x <= maximum))
 }
 
+# round + toSTring
+rst = function(v, digits = 3)
+  toString(round(v, digits))
+
 `%||%` = function(x, y) {
   if(is.null(x)) y else x
 }
@@ -112,7 +116,7 @@ disableMutationModels = function(x, disable, verbose = FALSE) {
 }
 
 
-# TODO: Move to pedtools
+# TODO: Move to pedtools?
 fixAllelesAndFreqs = function(alleles = NULL, afreq = NULL,
                               observed = NULL, NAstrings = c(0, "", NA, "-")) {