+++ NEWS.md
@@ -1,3 +1,15 @@
+# forrel 1.3.0
+
+# New features
+
+* The new function `ibdEstimate()` replaces the previous `IBDestimate()` (note the name change). This is a complete rewrite, which optimises the log-likelihood using a projected gradient descent algorithm, combined with a version of Armijo line search.
+
+* The function `ibdBootstrap()` replaces the previous `kappaBootstrap()` and `deltaBootstrap()`, and is considerably faster. This function implements both parametric and non-parametric bootstrap, controlled with the `method` parameter.
+
+* The output of `ibdEstimate()` now has a class attribute "ibdEst", with its own print and subsetting methods.
+
+
+
 # forrel 1.2.0
 
 ## New features
@@ -10,9 +22,9 @@
 
 ## Other changes
 
-* forrel now depends on pedtools version 0.9.6.
+* forrel now depends on version 0.9.6 (or later) of pedtools.
 
-* Deprecated arguments `id.labels`and `frametitles` in `missingPersonPlot()` has been removed.
+* Deprecated arguments `id.labels` and `frametitles` in `missingPersonPlot()` has been removed.
 
 
 # forrel 1.1.0